Quote pwd to account for directories with spaces
Turns out iCloud stores crap in a "Mobile Document" directory which
will confuse basename when it receives the `pwd` argument without quotes
because it thinks it's getting two arguments (one before the space, one after).

So we have to quote this expanded pwd output to insist that everything must
be fed to basename as a single argument.